+++ casbah-core/src/test/scala/MongoClientOptionsSpec.scala
@@ -35,7 +35,7 @@ class MongoClientOptionsSpec extends CasbahMutableSpecification {
 
     "have the same defaults as the Java MongoClientOptions" in {
 
-      val options = new MongoClientOptions.Builder().build
+      val options = MongoClientOptions()
       val javaOptions = new JavaMongoClientOptions.Builder().build
 
       options.getDescription must beEqualTo(javaOptions.getDescription)
@@ -51,6 +51,18 @@ class MongoClientOptionsSpec extends CasbahMutableSpecification {
       options.getSocketFactory must beEqualTo(javaOptions.getSocketFactory)
       options.getDbEncoderFactory must beEqualTo(javaOptions.getDbEncoderFactory)
       options.getDbDecoderFactory must beEqualTo(javaOptions.getDbDecoderFactory)
+
+      options.getAcceptableLatencyDifference must beEqualTo(javaOptions.getAcceptableLatencyDifference)
+      options.isAlwaysUseMBeans must beEqualTo(javaOptions.isAlwaysUseMBeans)
+      options.getHeartbeatConnectTimeout must beEqualTo(javaOptions.getHeartbeatConnectTimeout)
+      options.getHeartbeatFrequency must beEqualTo(javaOptions.getHeartbeatFrequency)
+      options.getHeartbeatSocketTimeout must beEqualTo(javaOptions.getHeartbeatSocketTimeout)
+      options.getHeartbeatThreadCount must beEqualTo(javaOptions.getHeartbeatThreadCount)
+      options.getMaxConnectionIdleTime must beEqualTo(javaOptions.getMaxConnectionIdleTime)
+      options.getMaxConnectionLifeTime must beEqualTo(javaOptions.getMaxConnectionLifeTime)
+      options.getMinConnectionsPerHost must beEqualTo(javaOptions.getMinConnectionsPerHost)
+      options.getHeartbeatConnectRetryFrequency must beEqualTo(javaOptions.getHeartbeatConnectRetryFrequency)
+      options.getRequiredReplicaSetName must beEqualTo(javaOptions.getRequiredReplicaSetName)
 
     }
 
@@ -81,6 +93,18 @@ class MongoClientOptionsSpec extends CasbahMutableSpecification {
       }
       builder.dbDecoderFactory(decoderFactory)
 
+      builder.acceptableLatencyDifference(100)
+      builder.alwaysUseMBeans(true)
+      builder.heartbeatConnectTimeout(100)
+      builder.heartbeatFrequency(100)
+      builder.heartbeatSocketTimeout(1000)
+      builder.heartbeatThreadCount(1)
+      builder.maxConnectionIdleTime(100)
+      builder.maxConnectionLifeTime(100)
+      builder.minConnectionsPerHost(100)
+      builder.heartbeatConnectRetryFrequency(200)
+      builder.requiredReplicaSetName("replicaSet")
+
       val options = builder.build
       options.getDescription must beEqualTo("test")
       options.getReadPreference must beEqualTo(ReadPreference.Secondary)
@@ -95,6 +119,18 @@ class MongoClientOptionsSpec extends CasbahMutableSpecification {
       options.getSocketFactory must beEqualTo(socketFactory)
       options.getDbEncoderFactory must beEqualTo(encoderFactory)
       options.getDbDecoderFactory must beEqualTo(decoderFactory)
+
+      options.getAcceptableLatencyDifference must beEqualTo(100)
+      options.isAlwaysUseMBeans must beEqualTo(true)
+      options.getHeartbeatConnectTimeout must beEqualTo(100)
+      options.getHeartbeatFrequency must beEqualTo(100)
+      options.getHeartbeatSocketTimeout must beEqualTo(1000)
+      options.getHeartbeatThreadCount must beEqualTo(1)
+      options.getMaxConnectionIdleTime must beEqualTo(100)
+      options.getMaxConnectionLifeTime must beEqualTo(100)
+      options.getMinConnectionsPerHost must beEqualTo(100)
+      options.getHeartbeatConnectRetryFrequency must beEqualTo(200)
+      options.getRequiredReplicaSetName must beEqualTo("replicaSet")
 
     }